+++ pkg/sdk/v0/gen/cmd/cli/root.go
@@ -20,7 +20,6 @@ func GenPluginRootCmd(gen *gen.Generator, sdkConfig *sdk.SdkConfig) error {
 	f.HeaderComment("generated by 'threeport-sdk gen' but will not be regenerated - intended for modification")
 
 	f.ImportAlias("github.com/threeport/threeport/pkg/cli/v0", "cli")
-	f.ImportAlias("github.com/threeport/threeport/pkg/config/v0", "config")
 
 	packageDir := strcase.ToSnake(sdkConfig.ModuleName)
 	commandLit := strcase.ToKebab(sdkConfig.ModuleName)
@@ -126,7 +125,7 @@ func GenPluginRootCmd(gen *gen.Generator, sdkConfig *sdk.SdkConfig) error {
 			Id("requestedControlPlane"),
 			Err(),
 		).Op(":=").Qual(
-			"github.com/threeport/threeport/pkg/config/v0",
+			"github.com/threeport/threeport/pkg/cli/v0",
 			"GetThreeportConfig",
 		).Call(Id("cliArgs").Dot("ControlPlaneName")),
 		If(Err().Op("!=").Nil()).Block(
